% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{annot.df}
\alias{annot.df}
\title{Example annotation data frame matching the example count matrix to 
test the expression curve estimation}
\format{
Data frame with three columns (cell, cell_type, donor)
}
\usage{
data(annotDfExample)
}
\description{
Contains three columns necessary for fitting the expression priors 
(cell barcode, cell type and donor)
}
\examples{
data(annotDfExample)
head(annot.df)
}
\keyword{datasets}
