% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{annot_smartseq}
\alias{annot_smartseq}
\title{Example Smart-seq2 annotation data frame (matching the Smart-seq2 count matrix)}
\format{
Annotation data frame with cell type and donor information per cell
}
\usage{
data(smartseqExample)
}
\description{
Contains an annotation data frame with cell type and donor information per cell
}
\examples{
data(smartseqExample)
head(annot_smartseq)
}
\keyword{datasets}
