% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{budget.optimization}
\alias{budget.optimization}
\title{Optimal parameters for increasing budget and simulated and observed priors from different
single cell technologies to reproduce figures curves in the vignette reproduce-figure-plots faster}
\format{Data frame with optimal parameter for increasing budget and different priors}
\usage{
data(precalculatedBudgetOptim)
}
\description{
Optimal parameters for increasing budget and simulated and observed priors from different
single cell technologies to reproduce figures curves in the vignette reproduce-figure-plots faster
}
\examples{
data(precalculatedBudgetOptim)
head(budget.optimization)
}
\keyword{datasets}
