% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{budgetCalculation.libPrepCell}
\alias{budgetCalculation.libPrepCell}
\title{Calculate total cost dependent on parameters for 10X design
using library preparation costs per cell}
\usage{
budgetCalculation.libPrepCell(
  nSamples,
  nCells,
  readDepth,
  prepCostCell,
  costFlowCell,
  readsPerFlowcell,
  rounding = FALSE
)
}
\arguments{
\item{nSamples}{Number of samples}

\item{nCells}{Cells per person}

\item{readDepth}{Read depth per cell}

\item{prepCostCell}{Library preparation costs per cell}

\item{costFlowCell}{Cost of one flow cells for sequencing}

\item{readsPerFlowcell}{Number reads that can be sequenced with one flow cell}

\item{rounding}{Rounds up the number of used kits and flow cells
(which might give a more realistic estimation of costs)}
}
\value{
Total experimental cost dependent on the parameters
}
\description{
Calculate total cost dependent on parameters for 10X design
using library preparation costs per cell
}
