% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{budgetCalculation}
\alias{budgetCalculation}
\title{Calculate total cost dependent on parameters}
\usage{
budgetCalculation(
  cellsPerPerson,
  readDepth,
  sampleSize,
  costKit,
  personsPerLane,
  costFlowCell,
  readsPerFlowcell,
  rounding = FALSE
)
}
\arguments{
\item{cellsPerPerson}{Cells per person}

\item{readDepth}{Read depth per cell}

\item{sampleSize}{Number of samples}

\item{costKit}{Cost for one 10X kit}

\item{personsPerLane}{Number of persons sequenced per lane}

\item{costFlowCell}{Cost of one flow cells for sequencing}

\item{readsPerFlowcell}{Number reads that can be sequenced with one flow cell}

\item{rounding}{Rounds up the number of used kits and flow cells
(which might give a more realistic estimation of costs)}
}
\value{
Total experimental cost dependent on the parameters
}
\description{
Calculate total cost dependent on parameters
}
