% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{budgetCalculation.restrictedDoublets}
\alias{budgetCalculation.restrictedDoublets}
\title{Calculate total cost dependent on parameters (adaptation with cells per lane instead of samples)}
\usage{
budgetCalculation.restrictedDoublets(
  nSamples,
  nCells,
  readDepth,
  costKit,
  cellsPerLane,
  costFlowCell,
  readsPerFlowcell,
  rounding = FALSE
)
}
\arguments{
\item{nSamples}{Number of samples}

\item{nCells}{Cells per individual}

\item{readDepth}{Read depth per cell}

\item{costKit}{Cost for one 10X kit}

\item{cellsPerLane}{Number of cells sequenced per lane}

\item{costFlowCell}{Cost of one flow cells for sequencing}

\item{readsPerFlowcell}{Number reads that can be sequenced with one flow cell}

\item{rounding}{Rounds up the number of used kits and flow cells
(which might give a more realistic estimation of costs)}
}
\value{
Total experimental cost dependent on the parameters
}
\description{
Calculate total cost dependent on parameters (adaptation with cells per lane instead of samples)
}
