% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_fit.R
\name{calculate.gene.counts}
\alias{calculate.gene.counts}
\title{Get expressed genes from pseudobulk}
\usage{
calculate.gene.counts(expr.array, min.counts = 10, perc.indiv = 0.5)
}
\arguments{
\item{expr.array}{3d pesudobulk matrix of genes x individuals x cell type (output of create.pseudobulk)}

\item{min.counts}{More than is number of UMI counts for each gene per person and cell type is required
to defined it as expressed in one individual}

\item{perc.indiv}{Percentage of individuals that need to have this gene expressed
to define it as globally expressed}
}
\value{
Data frame with all expressed genes per cell type (columns: gene name, cell type,
fraction in how many samples the gene is expressed)
}
\description{
Return a data frame with all genes, which are defined as expressed per cell type (at least min.counts
in perc.indiv fraction of all individuals)
}
