% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{cellsBudgetCalculation.libPrepCell}
\alias{cellsBudgetCalculation.libPrepCell}
\title{Estimate possible number of cells per individual depending on the total cost,
using library preparation costs per cell}
\usage{
cellsBudgetCalculation.libPrepCell(
  nSamples,
  readDepth,
  totalCost,
  prepCostCell,
  costFlowCell,
  readsPerFlowcell
)
}
\arguments{
\item{nSamples}{Number of samples}

\item{readDepth}{Read depth per cell}

\item{totalCost}{Experimental budget}

\item{prepCostCell}{Library preparation costs per cell}

\item{costFlowCell}{Cost of one flow cells for sequencing}

\item{readsPerFlowcell}{Number reads that can be sequenced with one flow cell}
}
\value{
Cells per individual that can be measured with this budget and other parameters
}
\description{
Estimate possible number of cells per individual depending on the total cost,
using library preparation costs per cell
}
