% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{cellsBudgetCalculation}
\alias{cellsBudgetCalculation}
\title{Estimate possible number of cells per individual depending on the total cost
and the other parameters for 10X design}
\usage{
cellsBudgetCalculation(
  nSamples,
  readDepth,
  totalCost,
  costKit,
  samplesPerLane,
  costFlowCell,
  readsPerFlowcell
)
}
\arguments{
\item{nSamples}{Number of samples}

\item{readDepth}{Read depth per cell}

\item{totalCost}{Experimental budget}

\item{costKit}{Cost for one 10X kit}

\item{samplesPerLane}{Number of individuals sequenced per lane}

\item{costFlowCell}{Cost of one flow cells for sequencing}

\item{readsPerFlowcell}{Number reads that can be sequenced with one flow cell}
}
\value{
Cells per individual that can be measured with this budget and other parameters
}
\description{
Estimate possible number of cells per individual depending on the total cost
and the other parameters for 10X design
}
