% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{cellsBudgetCalculation.restrictedDoublets}
\alias{cellsBudgetCalculation.restrictedDoublets}
\title{Estimate possible number of cells per individual depending on the total cost
and the other parameters (with a restricted number of cells per lane)}
\usage{
cellsBudgetCalculation.restrictedDoublets(
  nSamples,
  readDepth,
  totalCost,
  costKit,
  cellsPerLane,
  costFlowCell,
  readsPerFlowcell,
  reactionsPerKit = 6
)
}
\arguments{
\item{nSamples}{Number of samples}

\item{readDepth}{Read depth per cell}

\item{totalCost}{Experimental budget}

\item{costKit}{Cost for one 10X kit}

\item{cellsPerLane}{Number of cells sequenced per lane}

\item{costFlowCell}{Cost of one flow cells for sequencing}

\item{readsPerFlowcell}{Number reads that can be sequenced with one flow cell}

\item{reactionsPerKit}{Reactions (=lanes) per kit, defines the total
number of tested individuals per kit}
}
\value{
Cells per individual that can be measured with this budget and other parameters
}
\description{
Approximation without rounding
}
