% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{cellsPersonBudgetCalculation.restrictedDoublets}
\alias{cellsPersonBudgetCalculation.restrictedDoublets}
\title{Estimate possible number of cells per person depending on the total cost
and the other parameters (with a restricted number of cells per lane)}
\usage{
cellsPersonBudgetCalculation.restrictedDoublets(
  nSamples,
  readDepth,
  totalCost,
  costKit,
  cellsPerLane,
  costFlowCell,
  readsPerFlowcell
)
}
\arguments{
\item{nSamples}{Number of samples}

\item{readDepth}{Read depth per cell}

\item{totalCost}{Experimental budget}

\item{costKit}{Cost for one 10X kit}

\item{cellsPerLane}{Number of cells sequenced per lane}

\item{costFlowCell}{Cost of one flow cells for sequencing}

\item{readsPerFlowcell}{Number reads that can be sequenced with one flow cell}
}
\value{
Cells per person that can be measured with this budget and other parameters
}
\description{
Approximation without rounding
}
