% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_fit.R
\name{convert.gamma.parameters}
\alias{convert.gamma.parameters}
\title{Converts two different gamma parameterizations, one using rate
and shape (rateshape) and one using mean and sd (meansd)}
\usage{
convert.gamma.parameters(gamma.fits, type = "meansd")
}
\arguments{
\item{gamma.fits}{Parameters from the fitted gamma function (mixture of two gamma components),
either with mean and sd values or with rate and shape values}

\item{type}{Way of conversion, either to meansd or to rateshape variant}
}
\value{
Converted parameters of the gamma function
}
\description{
The distributions of the mean values are fitted using the rateshape
version, the parameterization over the UMI counts is done using the
meansd version.
}
