% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{count.matrix.example}
\alias{count.matrix.example}
\title{Example count matrices to test the expression curve estimation}
\format{List with four count matrices (complete and subsampled to 75\%, 50\% and 25\%)}
\usage{
data(countMatrixExample)
}
\description{
Small subset from the 10X PBMC data to visualize expression curve
estimation in the introduction vignette
}
\examples{
data(countMatrixExample)
head(count.matrix.example)
}
\keyword{datasets}
