% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{counts_smartseq}
\alias{counts_smartseq}
\title{Example Smart-seq2 count matrix for the tutorial}
\format{
Read count matrix of genes times cells
}
\usage{
data(smartseqExample)
}
\description{
Contains a read count matrix of genes times cells
}
\examples{
data(smartseqExample)
dim(counts_smartseq)
}
\keyword{datasets}
