% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_fit.R
\name{create.pseudobulk}
\alias{create.pseudobulk}
\title{Reformat count matrix to 3D pseudocount matrix}
\usage{
create.pseudobulk(expr.singlets, annotation, colName = "cell.type")
}
\arguments{
\item{expr.singlets}{gene count matrix (already filtered for multiplets etc.)}

\item{annotation}{Data frame with annotation of each cell to a cell type and donor,
the data frame must be sorted in the same way as the expression matrix (required columns: individual)}

\item{colName}{Column name of the cell type annotation in the annotation data frame}
}
\value{
3d matrix of genes x individuals x cell type
}
\description{
Sum up the counts per cell type and person for each gene and create a
3 dimensional matrix of genes x individuals x cell types
}
