% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{de.ref.study}
\alias{de.ref.study}
\title{DE priors}
\format{Data frame with prior information of DE genes}
\usage{
data(DERefStudy)
}
\description{
Data frame with DE priors (fold changes and gene expression ranks) for
calculation of the DE power. Taken from different celltype sorted bulk
DE studies.
}
\examples{
data(DERefStudy)
head(de.ref.study)
}
\keyword{datasets}
