% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{disp.fun.param.drop}
\alias{disp.fun.param.drop}
\title{Parameters for mean dispersion fit - Drop-seq data}
\format{
Data frame with parameter of mean-dispersion function
}
\usage{
data(dispFunParamDrop)
}
\description{
Parameters to estimate the dispersion parameter of each gene with the
mean of each gene, using the mean-dispersion function of DESeq
}
\examples{
data(dispFunParamDrop)
head(disp.fun.param.drop)
}
\keyword{datasets}
