% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{disp.fun.param}
\alias{disp.fun.param}
\title{Parameters for mean dispersion fit}
\format{
Data frame with parameter of mean-dispersion function
}
\usage{
data(dispFunParam)
}
\description{
Parameters to estimate the dispersion parameter of each gene with the
mean of each gene, using the mean-dispersion function of DESeq
}
\examples{
data(dispFunParam)
head(disp.fun.param)
}
\keyword{datasets}
