% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{disp.fun.param.smart}
\alias{disp.fun.param.smart}
\title{Parameters for mean dispersion fit - Smart-seq2 data}
\format{Data frame with parameter of mean-dispersion function}
\usage{
data(dispFunParamSmart)
}
\description{
Parameters to estimate the dispersion parameter of each gene
linear dependent on the read depth for the mean-dispersion function of DESeq
}
\examples{
data(dispFunParamSmart)
head(disp.fun.param.smart)
}
\keyword{datasets}
