% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_fit.R
\name{dispersion.function.estimation}
\alias{dispersion.function.estimation}
\title{Get median values of parameters from the mean-dispersion fits
(no correlation with UMI counts visible)}
\usage{
dispersion.function.estimation(disp.funs)
}
\arguments{
\item{disp.funs}{Data.frame with parameters of mean-dispersion fit
(asymptDisp, extraPois)}
}
\value{
Median parameters
}
\description{
Get median values of parameters from the mean-dispersion fits
(no correlation with UMI counts visible)
}
