% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{effectSize.DE.simulation}
\alias{effectSize.DE.simulation}
\title{Simulation of effect sizes for DE genes (FoldChanges)}
\usage{
effectSize.DE.simulation(mean, sd, numGenes)
}
\arguments{
\item{mean}{of the normal distribution of the log fold changes}

\item{sd}{of the normal distribution of the log fold changes}

\item{numGenes}{Number of effect sizes to simulate}
}
\value{
Vector with uniforml distributed gene ranks
}
\description{
Simulation of effect sizes for DE genes (FoldChanges)
}
