% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{eqtl.ref.study}
\alias{eqtl.ref.study}
\title{eQTL priors}
\format{
Data frame with prior information of eQTL genes
}
\usage{
data(eQTLRefStudy)
}
\description{
Data frame with eQTL priors (beta values and gene expression ranks) for
calculation of the eQTL power. Taken from different celltype sorted bulk
eQTL studies.
}
\examples{
data(eQTLRefStudy)
head(eqtl.ref.study)
}
\keyword{datasets}
