% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_fit.R
\name{estimate.exp.prob.param}
\alias{estimate.exp.prob.param}
\title{Estimate gene expression probability based on experimental parameters}
\usage{
estimate.exp.prob.param(
  nSamples,
  readDepth,
  nCellsCt,
  read.umi.fit,
  gamma.mixed.fits,
  gamma.probs,
  ct,
  disp.fun.param,
  min.counts = 10,
  perc.indiv = 0.5,
  nGenes = 21000,
  samplingMethod = "quantiles"
)
}
\arguments{
\item{nSamples}{Sample size}

\item{readDepth}{Target read depth per cell}

\item{nCellsCt}{Mean number of cells per person and cell type}

\item{read.umi.fit}{Data frame for fitting the mean UMI counts per cell depending on the mean readds per cell
(required columns: intercept, reads (slope))}

\item{gamma.mixed.fits}{Data frame with gamma mixed fit parameters for each cell type
(required columns: parameter, ct (cell type), intercept, meanUMI (slope))}

\item{gamma.probs}{Data frame with probability parameter of the gamma distributions
(required columns:ct (cell type), pi.c1 (probability of component 1))}

\item{ct}{Cell type of interest (name from the gamma mixed models)}

\item{disp.fun.param}{Function to fit the dispersion parameter dependent on the mean
(required columns: ct (cell type), asymptDisp, extraPois (both from taken from DEseq))}

\item{min.counts}{More than is number of UMI counts for each gene per person and cell type is required to defined it as expressed in one individual}

\item{perc.indiv}{Percentage of individuals that need to have this gene expressed to define it as globally expressed}

\item{nGenes}{Number of genes to simulate (should match the number of genes used for the fitting)}

\item{samplingMethod}{Approach to sample the gene mean values (either taking quantiles or random sampling)}
}
\value{
Vector with expression probabilities for each gene
}
\description{
This function estimates the expression probability of each gene in pseudobulk
with a certain cutoff of more than min.counts UMI counts, based on experimental
parameters which lead to certain mean and dispersion values for each gene
}
