% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_fit.R
\name{estimate.gene.counts}
\alias{estimate.gene.counts}
\title{Estimate gene expression probability}
\usage{
estimate.gene.counts(mu, size, cells.per.person, min.counts = 10,
  num.indivs = 14, perc.indiv = 0.5)
}
\arguments{
\item{mu}{Estimated mean value in sc data}
}
\description{
This function estimates the expression probability of each gene in pseudobulk
with a certain cutoff.
}
