% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_fit.R
\name{estimate.mean.dsp.values}
\alias{estimate.mean.dsp.values}
\title{Get the mean and dispersion values for each genes}
\usage{
estimate.mean.dsp.values(
  readDepth,
  read.umi.fit,
  gamma.mixed.fits,
  ct,
  disp.fun.param,
  nGenes = 21000,
  samplingMethod = "quantiles"
)
}
\arguments{
\item{readDepth}{Target read depth per cell}

\item{read.umi.fit}{Data frame for fitting the mean UMI counts per cell depending on the mean readds per cell
(required columns: intercept, reads (slope))}

\item{gamma.mixed.fits}{Data frame with gamma mixed fit parameters for each cell type
(required columns: parameter, ct (cell type), intercept, meanUMI (slope))}

\item{ct}{Cell type of interest (name from the gamma mixed models)}

\item{disp.fun.param}{Function to fit the dispersion parameter dependent on the mean
(required columns: ct (cell type), asymptDisp, extraPois (both from taken from DEseq))}

\item{nGenes}{Number of genes to simulate (should match the number of genes used for the fitting)}

\item{samplingMethod}{Approach to sample the gene mean values (either taking quantiles or random sampling)}
}
\value{
Data frame with mean and dispersion value for each gene
}
\description{
This function estimates the mean and dispersion of each gene (in a single cell,
before pseudobulk aggregation) dependent on the read depth and
the gamma mixture distribution
}
