% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{estimate.size.simulation}
\alias{estimate.size.simulation}
\title{Estimation of size parameter (1/dsp) for eQTL power simulation}
\usage{
estimate.size.simulation(count.mean, Rsq, af)
}
\arguments{
\item{count.mean}{Mean of negative binomial distribution for the counts}

\item{Rsq}{Effect size}

\item{af}{Allele frequency}
}
\value{
Numerical optimized size parameter (1/dispersion) to model the standard error
as good as possible in the simulation
}
\description{
Estimation of size parameter (1/dsp) for eQTL power simulation
}
