% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{fdr.optimization}
\alias{fdr.optimization}
\title{Function for numeric optimization to get an FDR corrected significance thresold}
\usage{
fdr.optimization(
  x,
  fdr,
  m0,
  type,
  exp.vector,
  es.vector,
  nSamples,
  mean.vector = NULL,
  disp.vector = NULL
)
}
\arguments{
\item{x}{Adjusted significance threshold to be found during optimization}

\item{fdr}{Chosen FDR threshold}

\item{m0}{Number of null hypothesis}

\item{type}{Either DE or eQTL power}

\item{exp.vector}{Vector of expression probabilities for each DEG/eQTL}

\item{es.vector}{Effect size vector (Fold Change for DEGs, Rsq for eQTLs)}

\item{nSamples}{Sample size}

\item{mean.vector}{Mean value for each DEG (only required for DE power)}

\item{disp.vector}{Dispersion value for each DEG (only required for DE power)}
}
\value{
Optimizing this function to 0 will lead to the correct adjusted
significance threshold x
}
\description{
Function for numeric optimization to get an FDR corrected significance thresold
}
