% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gamma.mixed.fits.drop}
\alias{gamma.mixed.fits.drop}
\title{Parameters for gamma umi fit - Drop-seq}
\format{Data frame with parameter of UMI-gamma parameter fit}
\usage{
data(gammaMixedFitsDrop)
}
\description{
Parameters to estimate parameter for the gamma mixed distribution of the
gene means with the mean UMI counts per cell using a linear function,
fitted from example lung Drop-seq data set
}
\examples{
data(gammaMixedFitsDrop)
head(gamma.mixed.fits.drop)
}
\keyword{datasets}
