% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gamma.mixed.fits}
\alias{gamma.mixed.fits}
\title{Parameters for gamma umi fit}
\format{Data frame with parameter of UMI-gamma parameter fit}
\usage{
data(gammaUmiFits)
}
\description{
Parameters to estimate parameter for the gamma mixed distribution of the
gene means with the mean UMI counts per cell using a linear function,
fitted from example PBMC 10X data set
}
\examples{
data(gammaUmiFits)
head(gamma.mixed.fits)
}
\keyword{datasets}
