% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gamma.mixed.fits.smart}
\alias{gamma.mixed.fits.smart}
\title{Parameters for gamma read fit - Smart-seq2 data}
\format{
Data frame with parameter of read-gamma parameter fit
}
\usage{
data(gammaMixedFitsSmart)
}
\description{
Parameters to estimate parameter for the gamma mixed distribution of the
gene means with the mean read counts per cell using a linear function,
fitted from example pancreas Smart-seq2 data
}
\examples{
data(gammaMixedFitsSmart)
head(gamma.mixed.fits.smart)
}
\keyword{datasets}
