% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gamma.probs.drop}
\alias{gamma.probs.drop}
\title{Probability parameter for the gamma umi fit - Drop-seq data}
\format{Data frame with probability parameter of UMI-gamma parameter fit}
\usage{
data(gammaProbsDrop)
}
\description{
Probability parameter of the first component of the gamma mixed model,
in contrast to the other parameters independent of the mean UMI counts per cell,
fitted from example PBMC 10X data set (mean value per cell type)
}
\examples{
data(gammaProbsDrop)
head(gamma.probs.drop)
}
\keyword{datasets}
