% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gamma.probs}
\alias{gamma.probs}
\title{Probability parameter for the gamma umi fit}
\format{Data frame with probability parameter of UMI-gamma parameter fit}
\usage{
data(gammaProbs)
}
\description{
Probability parameter of the first component of the gamma mixed model,
in contrast to the other parameters independent of the mean UMI counts per cell,
fitted from our example PBMC data set (mean value per cell type)
}
\examples{
data(gammaProbs)
head(gamma.probs)
}
\keyword{datasets}
