% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gamma.probs.smart}
\alias{gamma.probs.smart}
\title{Probability parameter for the gamma read fit - Smart-seq2 data}
\format{Data frame with probability parameter of read-gamma parameter fit}
\usage{
data(gammaProbs)
}
\description{
Probability parameter of the first component of the gamma mixed model,
in contrast to the other parameters independent of the mean read counts per cell,
fitted from pancreas Smart-seq2 data (mean value per cell type)
}
\examples{
data(gammaProbsSmart)
head(gamma.probs.smart)
}
\keyword{datasets}
