% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_fit.R
\name{gamma.umi.relation}
\alias{gamma.umi.relation}
\title{Estimate the relationship between the gamma mixed parameter and the mean UMI counts}
\usage{
\method{gamma}{umi.relation}()
}
\description{
Estimate the relationship between the gamma mixed parameter and the mean UMI counts
}
