% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{gene.rank.calculation}
\alias{gene.rank.calculation}
\title{Calculation of gene ranks from a normalized count matrix}
\usage{
gene.rank.calculation(countMatrix, diff.expr.genes)
}
\arguments{
\item{countMatrix}{normalized count matrix with gene names as row names}

\item{diff.expr.genes}{Vector of significant genes (DE or eQTLs), the gene names
in the vector need to match the row names of the count matrix}
}
\value{
Data frame with expression rank for each DE gene
}
\description{
Calculation of gene ranks from a normalized count matrix
}
