% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{gene.rank.calculation.vector}
\alias{gene.rank.calculation.vector}
\title{Calculation of gene ranks using a vector of mean expression per gene}
\usage{
gene.rank.calculation.vector(meanVector, geneNames, diff.expr.genes)
}
\arguments{
\item{meanVector}{Vector with mean value per gene in the sample
(use normalized counts)}

\item{geneNames}{Gene name corresponding to the mean value
(need to be in the same order as the mean vector)}

\item{diff.expr.genes}{Vector of significant genes (DE or eQTLs), the gene names
in the vector need to match the row names of the count matrix}
}
\value{
Data frame with expression rank for each DE gene
}
\description{
Calculation of gene ranks using a vector of mean expression per gene
}
