% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gene_length}
\alias{gene_length}
\title{Example gene length data frame}
\format{
Gene length data frame with gene symbol and gene length in bp (exonic regions)
}
\usage{
data(smartseqExample)
}
\description{
Contains gene length (only exonic regions) in bp for the genes from the Smart-seq2 example data frame;
information obtained from UCSC (gene annotation GRCh37hg19)
}
\examples{
data(smartseqExample)
head(gene_length)
}
\keyword{datasets}
