% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_fit.R
\name{meanUMI.calculation}
\alias{meanUMI.calculation}
\title{Calculate the mean UMI count per cell}
\usage{
meanUMI.calculation(countMatrix)
}
\arguments{
\item{countMatrix}{Raw UMI count matrix with genes as rows and cells as columns}
}
\value{
Mean UMI count per cell
}
\description{
Calculate the mean UMI count per cell
}
