% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_fit.R
\name{nbinom.estimation}
\alias{nbinom.estimation}
\title{Estimate the mean and dispersion parameter for each gene}
\usage{
nbinom.estimation(counts.ct)
}
\arguments{
\item{counts.ct}{Count matrix as in import for DEseq}
}
\value{
list with three data frame containing the normalized mean values,
the dispersion parameters and the parameters for the mean-dispersion function
estimated from DESeq
}
\description{
Use the DEseq together with the size factor normalization poscounts of DESeq2
}
