% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{number.cells.detect.celltype}
\alias{number.cells.detect.celltype}
\title{Cell sample size calculation for cell type identification}
\usage{
number.cells.detect.celltype(prob.cutoff, min.num.cells, cell.type.frac,
  num.indivs)
}
\arguments{
\item{prob.cutoff}{Target power to detect the cell type}

\item{min.num.cells}{Minimal number of the cells for the cell type of interest that should be detected}

\item{cell.type.frac}{Frequency of the cell type of interest}

\item{num.indivs}{Sample size (number of individuals)}
}
\value{
Required number of cells per person to reach the target power
}
\description{
Calculate required number of cells per person to detect at least min.num.cells of a cell type
with class frequency cell.type.frac for each of num.indivs individual
with a probability of prob.cutoff
}
