% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{observed.gene.counts}
\alias{observed.gene.counts}
\title{Observed gene counts from multiple data sets to reproduce figures curves in the vignette
reproduce-figure-plots faster}
\format{
Data frame with observed gene counts from multiple data sets
}
\usage{
data(precalculatedObservedGeneCounts)
}
\description{
Observed gene counts from multiple data sets to reproduce figures curves in the vignette
reproduce-figure-plots faster
}
\examples{
data(precalculatedObservedGeneCounts)
head(observed.gene.counts)
}
\keyword{datasets}
