% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{optimize.constant.budget.libPrepCell}
\alias{optimize.constant.budget.libPrepCell}
\title{Optimizing cost parameters to maximize detection power for a given budget with
library preparation costs per cell}
\usage{
optimize.constant.budget.libPrepCell(
  totalBudget,
  type,
  ct,
  ct.freq,
  prepCostCell,
  costFlowCell,
  readsPerFlowcell,
  ref.study,
  ref.study.name,
  samplesPerLane,
  read.umi.fit,
  gamma.mixed.fits,
  disp.fun.param,
  nSamplesRange = NULL,
  nCellsRange = NULL,
  readDepthRange = NULL,
  mappingEfficiency = 0.8,
  multipletRate = 7.67e-06,
  multipletFactor = 1.82,
  min.UMI.counts = 3,
  perc.indiv.expr = 0.5,
  nGenes = 21000,
  samplingMethod = "quantiles",
  multipletRateGrowth = "linear"
)
}
\arguments{
\item{totalBudget}{Overall experimental budget}

\item{type}{(eqtl/de) study}

\item{ct}{Cell type of interest (name should be found in the gamma and dispersion models)}

\item{ct.freq}{Frequency of the cell type of interest}

\item{prepCostCell}{Library preparation costs per cell}

\item{costFlowCell}{Cost of one flow cells for sequencing}

\item{readsPerFlowcell}{Number reads that can be sequenced with one flow cell}

\item{ref.study}{Data frame with reference studies to be used for effect sizes and ranks
(required columns: name (study name), rank (expression rank), FoldChange (DE study) /Rsq (eQTL study))}

\item{ref.study.name}{Name of the reference study. Will be checked in the ref.study data frame for it (as column name).}

\item{samplesPerLane}{Maximal number of individuals per 10X lane}

\item{read.umi.fit}{Data frame for fitting the mean UMI counts per cell depending on the mean readds per cell
(required columns: intercept, reads (slope))}

\item{gamma.mixed.fits}{Data frame with gamma mixed fit parameters for each cell type
(required columns: parameter, ct (cell type), intercept, meanUMI (slope))}

\item{disp.fun.param}{Function to fit the dispersion parameter dependent on the mean
(required columns: ct (cell type), asymptDisp, extraPois (both from taken from DEseq))}

\item{nSamplesRange}{Range of sample sizes that should be tested (vector)}

\item{nCellsRange}{Range of cells per individual that should be tested (vector)}

\item{readDepthRange}{Range of read depth values that should be tested (vector)}

\item{mappingEfficiency}{Fraction of reads successfully mapped to the transcriptome in the end (need to be between 1-0)}

\item{multipletRate}{Expected increase in multiplets for additional cell in the lane}

\item{multipletFactor}{Expected read proportion of multiplet cells vs singlet cells}

\item{min.UMI.counts}{Expression defition parameter: more than is number of UMI counts for each
gene per individual and cell type is required to defined it as expressed in one individual}

\item{perc.indiv.expr}{Expression defition parameter: percentage of individuals that need to have this gene expressed
to define it as globally expressed}

\item{nGenes}{Number of genes to simulate (should match the number of genes used for the fitting)}

\item{samplingMethod}{Approach to sample the gene mean values (either taking quantiles or random sampling)}

\item{multipletRateGrowth}{Development of multiplet rate with increasing number of cells per lane, "linear" if overloading should be
modeled explicitly, otherwise "constant". The default value for the parameter multipletRate is matching the option "linear".}
}
\description{
This function determines the optimal parameter combination for a given budget.
The optimal combination is thereby the one with the highest detection power.
Of the three parameters sample size, cells per sample and read depth, two need to be set and
the third one is uniquely defined given the other two parameters and the overall budget.
}
