% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{optimize.constant.budget.smartseq}
\alias{optimize.constant.budget.smartseq}
\title{Optimizing cost parameters to maximize detection power for a given budget and Smart-seq design}
\usage{
optimize.constant.budget.smartseq(
  totalBudget,
  type,
  ct,
  ct.freq,
  prepCostCell,
  costFlowCell,
  readsPerFlowcell,
  ref.study,
  ref.study.name,
  gamma.mixed.fits,
  disp.linear.fit,
  nSamplesRange = NULL,
  nCellsRange = NULL,
  readDepthRange = NULL,
  mappingEfficiency = 0.8,
  multipletFraction = 0,
  multipletFactor = 1.82,
  min.norm.count = 3,
  perc.indiv.expr = 0.5,
  nGenes = 21000,
  samplingMethod = "quantiles",
  sign.threshold = 0.05,
  MTmethod = "Bonferroni",
  useSimulatedPower = TRUE,
  speedPowerCalc = FALSE
)
}
\arguments{
\item{totalBudget}{Overall experimental budget}

\item{type}{(eqtl/de) study}

\item{ct}{Cell type of interest (name from the gamma mixed models)}

\item{ct.freq}{Frequency of the cell type of interest}

\item{prepCostCell}{Library preparation costs per cell}

\item{costFlowCell}{Cost of one flow cells for sequencing}

\item{readsPerFlowcell}{Number reads that can be sequenced with one flow cell}

\item{ref.study}{Data frame with reference studies to be used for expression ranks and effect sizes
(required columns: name (study name), rank (expression rank), FoldChange (DE study) /Rsq (eQTL study))}

\item{ref.study.name}{Name of the reference study. Will be checked in the ref.study data frame for it (as column name).}

\item{gamma.mixed.fits}{Data frame with gamma mixed fit parameters for each cell type
(required columns: parameter, ct (cell type), intercept, meanReads (slope))}

\item{disp.linear.fit}{Function to fit the dispersion parameter dependent on the mean (parameter linear dependent on read depth)
(required columns: parameter, ct (cell type), intercept, meanReads (slope))}

\item{nSamplesRange}{Range of sample sizes that should be tested (vector)}

\item{nCellsRange}{Range of cells per individual that should be tested (vector)}

\item{readDepthRange}{Range of read depth values that should be tested (vector)}

\item{mappingEfficiency}{Fraction of reads successfully mapped to the transcriptome in the end (need to be between 1-0)}

\item{multipletFraction}{Multiplet fraction in the experiment as a constant factor}

\item{multipletFactor}{Expected read proportion of multiplet cells vs singlet cells}

\item{min.norm.count}{Expression defition parameter: more than is number of counts per kilobase transcript for each
gene per individual and cell type is required to defined it as expressed in one individual}

\item{perc.indiv.expr}{Expression defition parameter: percentage of individuals that need to have this gene expressed
to define it as globally expressed}

\item{nGenes}{Number of genes to simulate (should match the number of genes used for the fitting)}

\item{samplingMethod}{Approach to sample the gene mean values (either taking quantiles or random sampling)}

\item{sign.threshold}{Significance threshold}

\item{MTmethod}{Multiple testing correction method (possible options: "Bonferroni","FDR","none")}

\item{useSimulatedPower}{Option to simulate eQTL power for small mean values to increase accuracy
(only possible for eQTL analysis)}

\item{speedPowerCalc}{Option to speed power calculation by skipping all genes with
an expression probability less than 0.01 (as overall power is anyway close to 0)}
}
\value{
Data frame with overall detection power, power and expression power for
each possible parameter combination given the budget and the parameter ranges
}
\description{
This function determines the optimal parameter combination for a given budget.
The optimal combination is thereby the one with the highest detection power.
Of the three parameters sample size, cells per sample and read depth, two need to be set and
the third one is uniquely defined given the other two parameters and the overall budget.
}
