% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{optimizeSizeFactor}
\alias{optimizeSizeFactor}
\title{Function for numeric optimization of size parameter}
\usage{
optimizeSizeFactor(x, sd.error, beta.mean)
}
\arguments{
\item{x}{Tested size factor}

\item{sd.error}{The targeted standard error}

\item{beta.mean}{The mean of the counts}
}
\value{
A value close to 0 shows a good accordance of targeted and simulated
standard error
}
\description{
Function for numeric optimization of size parameter
}
