% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_fit.R
\name{parametricDispersionFit_DEseq}
\alias{parametricDispersionFit_DEseq}
\title{Reimplementation of DEseq function for fitting mean-dispersion curve}
\usage{
parametricDispersionFit_DEseq(means, disps)
}
\arguments{
\item{means}{Mean vector for all expressed genes}

\item{disp}{Dispersion vector for all expressed genes}
}
\value{
Parameters of fitted mean-dispersion curve
}
\description{
Original code: Bioconductor R package DESeq, Simon Anders (EMBL Heidelberg),
DOI 10.18129/B9.bioc.DESeq
}
