% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power.eqtl}
\alias{power.eqtl}
\title{Wrapper funtion to use either simulated power or power based on F-test
(dependent on pseudobulk mean and used parameters)}
\usage{
power.eqtl(
  count.mean,
  heritability,
  sig.level,
  nSamples,
  useSimulatedPower = TRUE,
  simThreshold = 4
)
}
\arguments{
\item{count.mean}{Expression mean in the pseudobulk}

\item{heritability}{Heritability of the trait}

\item{sig.level}{Significane threshold}

\item{nSamples}{Sample size}

\item{useSimulatedPower}{Option to simulate eQTL power for small mean values
to increase accuracy}

\item{simThreshold}{Threshold until which the simulated power is taken
instead of the analytic}
}
\value{
Power to detect the eQTL gene
}
\description{
Wrapper funtion to use either simulated power or power based on F-test
(dependent on pseudobulk mean and used parameters)
}
