% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power.eqtl.simulated.help}
\alias{power.eqtl.simulated.help}
\title{Helper function for eQTL simulation power calculation}
\usage{
power.eqtl.simulated.help(count.mean, Rsq, nSamples, size.estimates, af = NULL)
}
\arguments{
\item{count.mean}{Expression mean in the pseudobulk}

\item{Rsq}{Heritability of the trait}

\item{nSamples}{Sample size}

\item{size.estimates}{Data frame with precalculated size values to speed calculation}

\item{af}{Allele frequency (sampled if not explicity given)}
}
\description{
Helper function for eQTL simulation power calculation
}
