% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power.general.withDoublets}
\alias{power.general.withDoublets}
\title{Power calculation for a DE/eQTL study}
\usage{
power.general.withDoublets(nSamples, nCells, readDepth, ct.freq, type,
  ref.study, ref.study.name, personsPerLane, read.umi.fit,
  gamma.mixed.fits, multipletRate, multipletFactor, min.UMI.counts = 10,
  perc.indiv.expr = 0.5)
}
\arguments{
\item{nSamples}{Sample size}

\item{nCells}{Number of cells per person}

\item{readDepth}{Target read depth per cell}

\item{ct.freq}{Frequency of the cell type of interest}

\item{type}{(eqtl/de) study}

\item{ref.study}{Data frame with reference studies to be used for effect sizes and ranks (required columns:)}

\item{ref.study.name}{Name of the reference study. Will be checked in the ref.study data frame for it (as column name).}

\item{personsPerLane}{Maximal number of persons per 10X lane}
}
\value{
Power to detect the DE/eQTL genes from the reference study in a single cell experiment with these parameters
}
\description{
This function ...
}
