% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power.sameReadDepth.restrictedDoublets}
\alias{power.sameReadDepth.restrictedDoublets}
\title{Power calculation for a DE/eQTL study with same read depth as the fitted gamma distribution}
\usage{
power.sameReadDepth.restrictedDoublets(
  nSamples,
  nCells,
  ct.freq,
  type,
  ref.study,
  ref.study.name,
  cellsPerLane,
  gamma.parameters,
  ct,
  disp.fun.param,
  mappingEfficiency = 0.8,
  multipletRate = 7.67e-06,
  multipletFactor = 1.82,
  min.UMI.counts = 3,
  perc.indiv.expr = 0.5,
  nGenes = 21000,
  samplingMethod = "quantiles",
  multipletRateGrowth = "linear"
)
}
\arguments{
\item{nSamples}{Sample size}

\item{nCells}{Number of cells per individual}

\item{ct.freq}{Frequency of the cell type of interest}

\item{type}{(eqtl/de) study}

\item{ref.study}{Data frame with reference studies to be used for expression ranks and effect sizes
(required columns: name (study name), rank (expression rank), FoldChange (DE study) /Rsq (eQTL study))}

\item{ref.study.name}{Name of the reference study. Will be checked in the ref.study data frame for it (as column name).}

\item{cellsPerLane}{Maximal number of cells per 10X lane}

\item{gamma.parameters}{Data frame with gamma parameters for each cell type
(required columns: ct (cell type), s1, r1, s2, r2, p1, p2/p3 (gamma parameters for both components))}

\item{ct}{Cell type of interest (name from the gamma mixed models)}

\item{disp.fun.param}{Function to fit the dispersion parameter dependent on the mean
(required columns: ct (cell type), asymptDisp, extraPois (both from taken from DEseq))}

\item{mappingEfficiency}{Fraction of reads successfully mapped to the transcriptome in the end (need to be between 1-0)}

\item{multipletRate}{Expected increase in multiplets for additional cell in the lane}

\item{multipletFactor}{Expected read proportion of multiplet cells vs singlet cells}

\item{min.UMI.counts}{Expression defition parameter: more than is number of UMI counts for each
gene per individual and cell type is required to defined it as expressed in one individual}

\item{perc.indiv.expr}{Expression defition parameter: percentage of individuals that need to have this gene expressed
to define it as globally expressed}

\item{nGenes}{Number of genes to simulate (should match the number of genes used for the fitting)}

\item{samplingMethod}{Approach to sample the gene mean values (either taking quantiles or random sampling)}

\item{multipletRateGrowth}{Development of multiplet rate with increasing number of cells per lane, "linear" if overloading should be
modeled explicitly, otherwise "constant". The default value for the parameter multipletRate is matching the option "linear".}
}
\value{
Power to detect the DE/eQTL genes from the reference study in a single cell experiment with these parameters
}
\description{
This function is a variant of power.sameReadDepth.withDoublets, where not the number of samplesPerLane is given as an
parameter, but instead the individuals are distributed over the lanes in a way that restricts the total number of
cells per lane instead. This gives also an upper bound for the doublet rate.
}
