% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power.smartseq}
\alias{power.smartseq}
\title{Power calculation for a DE/eQTL study with Smart-seq design}
\usage{
power.smartseq(
  nSamples,
  nCells,
  readDepth,
  ct.freq,
  type,
  ref.study,
  ref.study.name,
  gamma.mixed.fits,
  ct,
  disp.linear.fit,
  mappingEfficiency = 0.8,
  multipletFraction = 0,
  multipletFactor = 1.82,
  min.norm.count = 3,
  perc.indiv.expr = 0.5,
  cutoffVersion = "absolute",
  nGenes = 21000,
  samplingMethod = "quantiles",
  sign.threshold = 0.05,
  MTmethod = "Bonferroni",
  useSimulatedPower = TRUE,
  simThreshold = 4,
  speedPowerCalc = FALSE,
  indepSNPs = 10,
  ssize.ratio.de = 1,
  returnResultsDetailed = FALSE
)
}
\arguments{
\item{nSamples}{Sample size}

\item{nCells}{Number of cells per individual}

\item{readDepth}{Target read depth per cell}

\item{ct.freq}{Frequency of the cell type of interest}

\item{type}{(eqtl/de) study}

\item{ref.study}{Data frame with reference studies to be used for expression ranks and effect sizes
(required columns: name (study name), rank (expression rank), FoldChange (DE study) /Rsq (eQTL study))}

\item{ref.study.name}{Name of the reference study. Will be checked in the ref.study data frame for it (as column name).}

\item{gamma.mixed.fits}{Data frame with gamma mixed fit parameters for each cell type
(required columns: parameter, ct (cell type), intercept, meanReads (slope))}

\item{ct}{Cell type of interest (name from the gamma mixed models)}

\item{disp.linear.fit}{Function to fit the dispersion parameter dependent on the mean (parameter linear dependent on read depth)
(required columns: parameter, ct (cell type), intercept, meanReads (slope))}

\item{mappingEfficiency}{Fraction of reads successfully mapped to the transcriptome
in the end (need to be between 1-0)}

\item{multipletFraction}{Multiplet fraction in the experiment as a constant factor}

\item{multipletFactor}{Expected read proportion of multiplet cells vs singlet cells}

\item{cutoffVersion}{Either "absolute" or "percentage" leading to different
interpretations of min.counts (see description above)}

\item{nGenes}{Number of genes to simulate (should match the number of genes used for the fitting)}

\item{samplingMethod}{Approach to sample the gene mean values (either taking quantiles
or random sampling)}

\item{sign.threshold}{Significance threshold}

\item{MTmethod}{Multiple testing correction method (possible options: "Bonferroni","FDR","none")}

\item{useSimulatedPower}{Option to simulate eQTL power for small mean values to increase accuracy
(only possible for eQTL analysis)}

\item{simThreshold}{Threshold until which the simulated power is taken instead of the analytic}

\item{speedPowerCalc}{Option to speed power calculation by skipping all genes with
an expression probability less than 0.01 (as overall power is anyway close to 0)}

\item{indepSNPs}{Number of independent SNPs assumed for each loci (for eQTL
Bonferroni multiple testing correction the number of tests are estimated
as number expressed genes * indepSNPs)}

\item{ssize.ratio.de}{In the DE case, ratio between sample size of group 0
(control group) and group 1 (1=balanced design)}

\item{returnResultsDetailed}{If true, return not only summary data frame, but additional list with exact probability vectors}
}
\value{
Power to detect the DE/eQTL genes from the reference study in a single cell experiment with these parameters
}
\description{
This function to calculate the detection power for a DE or eQTL study, given DE/eQTL genes from a reference study,
in a single cell Smart-seq RNAseq study. The power depends on the cost determining parameter of sample size, number of cells
per individual and read depth.
}
