% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{print_optimalDesign_10X}
\alias{print_optimalDesign_10X}
\title{Print design parameters of optimal design}
\usage{
print_optimalDesign_10X(
  optim_output,
  samplesPerLane,
  readsPerFlowcell,
  reactionsPerKit = 6
)
}
\arguments{
\item{optim_output}{Result data frame of the optimization functions
(optimize.constant.budget and variants of the function)}

\item{samplesPerLane}{Number of individuals sequenced per lane}

\item{readsPerFlowcell}{Number reads that can be sequenced with one flow cell}

\item{reactionsPerKit}{Reactions (=lanes) per kit, defines the total
number of tested individuals per kit}
}
\value{
Nothing (print output instead)
}
\description{
Convenience function to read the output of the budget optimization functions
more easily: selected the row with the optimal detection power and showing
sample size, cells and read depth as well as the number of 10X kits and flow cells
for this specific design
}
