% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{read.umi.fit}
\alias{read.umi.fit}
\title{Parameters for read umi fit}
\format{
Data frame with parameter of read-UMI fit
}
\usage{
data(readDepthUmiFit)
}
\description{
Parameters to estimate the mean UMI counts per cell with the mean read depth,
using a logarithmic function, fitted from our example PBMC data set
}
\examples{
data(readDepthUmiFit)
head(read.umi.fit)
}
\keyword{datasets}
