% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{readDepthBudgetCalculation.libPrepCell}
\alias{readDepthBudgetCalculation.libPrepCell}
\title{Estimate possible read depth depending on the total cost,
using library preparation costs per cell}
\usage{
readDepthBudgetCalculation.libPrepCell(
  nSamples,
  nCells,
  totalCost,
  prepCostCell,
  costFlowCell,
  readsPerFlowcell
)
}
\arguments{
\item{nSamples}{Number of samples}

\item{nCells}{Cells per person}

\item{totalCost}{Experimental budget}

\item{prepCostCell}{Library preparation costs per cell}

\item{costFlowCell}{Cost of one flow cells for sequencing}

\item{readsPerFlowcell}{Number reads that can be sequenced with one flow cell}
}
\value{
Read depth that can be measured with this budget and other parameters
}
\description{
Estimate possible read depth depending on the total cost,
using library preparation costs per cell
}
