% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{readDepthBudgetCalculation}
\alias{readDepthBudgetCalculation}
\title{Estimate possible read depth depending on the total cost
and the other parameters for 10X design}
\usage{
readDepthBudgetCalculation(
  nSamples,
  nCells,
  totalCost,
  costKit,
  samplesPerLane,
  costFlowCell,
  readsPerFlowcell,
  reactionsPerKit = 6
)
}
\arguments{
\item{nSamples}{Number of samples}

\item{nCells}{Cells per individual}

\item{totalCost}{Experimental budget}

\item{costKit}{Cost for one 10X kit}

\item{samplesPerLane}{Number of individuals sequenced per lane}

\item{costFlowCell}{Cost of one flow cells for sequencing}

\item{readsPerFlowcell}{Number reads that can be sequenced with one flow cell}

\item{reactionsPerKit}{Reactions (=lanes) per kit, defines the total
number of tested individuals per kit}
}
\value{
Read depth that can be measured with this budget and other parameters
}
\description{
Estimate possible read depth depending on the total cost
and the other parameters for 10X design
}
