% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{readDepthBudgetCalculation.restrictedDoublets}
\alias{readDepthBudgetCalculation.restrictedDoublets}
\title{Estimate possible read depth depending on the total cost
and the other parameters (with a restricted number of cells per lane)}
\usage{
readDepthBudgetCalculation.restrictedDoublets(
  nSamples,
  nCells,
  totalCost,
  costKit,
  cellsPerLane,
  costFlowCell,
  readsPerFlowcell
)
}
\arguments{
\item{nSamples}{Number of samples}

\item{nCells}{Cells per individual}

\item{totalCost}{Experimental budget}

\item{costKit}{Cost for one 10X kit}

\item{cellsPerLane}{Number of cells sequenced per lane}

\item{costFlowCell}{Cost of one flow cells for sequencing}

\item{readsPerFlowcell}{Number reads that can be sequenced with one flow cell}
}
\value{
Read depth that can be measured with this budget and other parameters
}
\description{
Estimate possible read depth depending on the total cost
and the other parameters (with a restricted number of cells per lane)
}
