% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_fit.R
\name{sample.disp.values}
\alias{sample.disp.values}
\title{Sample the dispersion values dependent on mean values using the DESeq function parametrization}
\usage{
sample.disp.values(mean.vals, disp.parameter)
}
\arguments{
\item{mean.vals}{Vector of gene means}

\item{disp.parameter}{Data frame with parameter of mean-dispersion function}
}
\value{
Vector with simulated mean values
}
\description{
WARNING: The used package might be changed later
}
