% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_fit.R
\name{sample.mean.values.quantiles}
\alias{sample.mean.values.quantiles}
\title{Draw the mean values using the two quantile distributions of the gamma mixed distribution}
\usage{
sample.mean.values.quantiles(gamma.parameters, nGenes = 21000)
}
\arguments{
\item{gamma.parameters}{Data frame with gamma parameters
(fitted using mixed.gamma.estimation)}

\item{nGenes}{Number of genes to sample}
}
\value{
Vector with simulated mean values
}
\description{
Draw the mean values using the two quantile distributions of the gamma mixed distribution
}
