% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_fit.R
\name{sample.mean.values.random}
\alias{sample.mean.values.random}
\title{Randomly sample the mean values using a gamma mixed distribution}
\usage{
sample.mean.values.random(gamma.parameters, nGenes = 21000)
}
\arguments{
\item{gamma.parameters}{Data frame with gamma parameters
(fitted using mixed.gamma.estimation)}

\item{nGenes}{Number of genes to sample}
}
\value{
Vector with simulated mean values
}
\description{
WARNING: The used package might be changed later
}
