% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{sampleSizeBudgetCalculation.libPrepCell}
\alias{sampleSizeBudgetCalculation.libPrepCell}
\title{Estimate possible sample size depending on the total cost,
using library preparation costs per cell}
\usage{
sampleSizeBudgetCalculation.libPrepCell(
  nCells,
  readDepth,
  totalCost,
  prepCostCell,
  costFlowCell,
  readsPerFlowcell
)
}
\arguments{
\item{nCells}{Cells per individual}

\item{readDepth}{Read depth per cell}

\item{totalCost}{Experimental budget}

\item{prepCostCell}{Library preparation costs per cell}

\item{costFlowCell}{Cost of one flow cells for sequencing}

\item{readsPerFlowcell}{Number reads that can be sequenced with one flow cell}
}
\value{
Number of samples that can be measured with this budget and other parameters
}
\description{
A balanced design with two classes is assumed for the sample size calculation.
}
