% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{select.cutoffs}
\alias{select.cutoffs}
\title{Identifying the expression threshold combination that maximizes the detection power}
\usage{
select.cutoffs(
  umi_range,
  pop_range,
  nSamples,
  nCells,
  readDepth,
  ct.freq,
  type,
  ref.study,
  ref.study.name,
  cellsPerLane,
  read.umi.fit,
  gamma.mixed.fits,
  ct,
  disp.fun.param,
  ...
)
}
\arguments{
\item{umi_range}{Vector with UMI counts to test}

\item{pop_range}{Vector with population thresholds to test}

\item{nSamples}{Sample size}

\item{nCells}{Number of cells per individual}

\item{readDepth}{Target read depth per cell}

\item{ct.freq}{Frequency of the cell type of interest}

\item{type}{(eqtl/de) study}

\item{ref.study}{Data frame with reference studies to be used for expression ranks and effect sizes
(required columns: name (study name), rank (expression rank), FoldChange (DE study) /Rsq (eQTL study))}

\item{ref.study.name}{Name of the reference study. Will be checked in the ref.study data frame for it (as column name).}

\item{cellsPerLane}{Maximal number of cells per 10X lane}

\item{read.umi.fit}{Data frame for fitting the mean UMI counts per cell
depending on the mean readds per cell (required columns: intercept, reads (slope))}

\item{gamma.mixed.fits}{Data frame with gamma mixed fit parameters for each cell type
(required columns: parameter, ct (cell type), intercept, meanUMI (slope))}

\item{ct}{Cell type of interest (name from the gamma mixed models)}

\item{disp.fun.param}{Function to fit the dispersion parameter dependent on the mean
(required columns: ct (cell type), asymptDisp, extraPois (both from taken from DEseq))}

\item{...}{additional arguments that can be passed to
\code{\link[scPower:power.general.restrictedDoublets]{power.general.restrictedDoublets()}}
(excluding min.UMI.counts
and perc.indiv.expr that will be evaluated in the function)}
}
\value{
Results for threshold combination with the maximal detection power
}
\description{
Identifying the expression threshold combination that maximizes the detection power
}
