% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sim.eqtl.pvals}
\alias{sim.eqtl.pvals}
\title{Precalcuated p-values from eQTL simulation}
\format{
Matrix of with 100 p-values (columns) and rownames in the format
"mean_Rsq_sampleSize" showing the corresponding parameter combination with
which the p-values were generated
}
\usage{
data(simEqtlPvals)
}
\description{
Precalculated p-values from eQTL simulation to speed calculation time
}
\examples{
data(simEqtlPvals)
head(sim.eqtl.pvals)
}
\keyword{datasets}
