% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{simulated.DE.values}
\alias{simulated.DE.values}
\title{Results of DE power generated by simulations with powsimR}
\format{Data frame with simulated DE power}
\usage{
data(simulatedDEValues)
}
\description{
Results of DE power generated by simulations with powsimR
}
\examples{
data(simulatedDEValues)
head(simulated.DE.values)
}
\keyword{datasets}
