% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{simulated.eQTL.values}
\alias{simulated.eQTL.values}
\title{Results of eQTL power generated by self-implemented simulations
(see vignette reproduce-paper-plots)}
\format{
Data frame with simulated eQTL power
}
\usage{
data(simulatedEQTLValues)
}
\description{
Results of eQTL power generated by self-implemented simulations
(see vignette reproduce-paper-plots)
}
\examples{
data(simulatedEQTLValues)
head(simulated.eQTL.values)
}
\keyword{datasets}
