% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_fit.R
\name{sizeFactorsPosCounts}
\alias{sizeFactorsPosCounts}
\title{Reimplementation of DEseq2 function for size factors with option poscounts}
\usage{
sizeFactorsPosCounts(counts)
}
\arguments{
\item{counts}{UMI count matrix}
}
\value{
size factors to normalize each cell
}
\description{
Alternative to the standard size factor, better suited for scRNAseq data
with a lot of zero values
}
