% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_fit.R
\name{sizeFactorsStandard}
\alias{sizeFactorsStandard}
\title{Reimplementation of DEseq function for size factors}
\usage{
sizeFactorsStandard(counts)
}
\arguments{
\item{counts}{UMI count matrix}
}
\value{
size factors to normalize each cell
}
\description{
Original code: Bioconductor R package DESeq, Simon Anders (EMBL Heidelberg),
DOI 10.18129/B9.bioc.DESeq
}
