% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_fit.R
\name{umi.gamma.relation}
\alias{umi.gamma.relation}
\title{Estimate linear relationship between the gamma mixed parameter and the mean UMI counts}
\usage{
umi.gamma.relation(gamma.fits)
}
\arguments{
\item{gamma.fits}{Parameters from the fitted gamma function (pi.c1,pi.c2,mu.c1,mu.c2,sd.c1,sd.c2),
calculated in mixed.gamma.estimation, and a column with mean UMI values, can be calculated in
mean.umi.calculation}
}
\value{
List with linear fit for gamma parameters mean and sd and mean values for gamma parameters probability
}
\description{
Estimate linear relationship between the gamma mixed parameter and the mean UMI counts
}
