% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_fit.R
\name{umi.read.relation}
\alias{umi.read.relation}
\title{Estimate the relationship between mean UMI counts and mean read counts}
\usage{
umi.read.relation()
}
\description{
Estimate the relationship between mean UMI counts and mean read counts
}
