% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{uniform.ranks.interval}
\alias{uniform.ranks.interval}
\title{Simulation of gene ranks uniformly distributed in a certain interval}
\usage{
uniform.ranks.interval(start, end, numGenes)
}
\arguments{
\item{start}{Start of the rank interval (and value of the first simulated rank)}

\item{end}{End of the rank interval}

\item{numGenes}{Number of ranks to simulate}
}
\value{
Vector with uniforml distributed gene ranks
}
\description{
If the given interval is not completely dividable by the number
of genes to simulate (numGenes), the number of simulated ranks might
be slightly different from the numGenes.
}
