% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{visualize.gamma.fits}
\alias{visualize.gamma.fits}
\title{Plotting function to visualize simulated mean values from gamma distribution
compare to mean values used for the fitting}
\usage{
visualize.gamma.fits(
  mean.vals,
  gamma.parameters,
  nGenes = 21000,
  lower.dist = 1e-05,
  zero.pos = 1e-06,
  num.bins = 30
)
}
\arguments{
\item{mean.vals}{Vector with mean values used for fitting the gamma distribution}

\item{gamma.parameters}{Parameters of the gamma distribution
(already filtered for the cell type of interest)}

\item{nGenes}{number of genes used for fitting the gamma function
(parameter num.genes.kept in the function mixed.gamma.estimation)}

\item{lower.dist}{Lowest value which should be shown in the histogram}

\item{zero.pos}{Position where the zero component should be plotted at
(needs to be smaller than lower.dist)}

\item{num.bins}{Number of bins in the histogram}
}
\value{
ggplot object of a histogram with logarithmized x axis
}
\description{
Plotting function to visualize simulated mean values from gamma distribution
compare to mean values used for the fitting
}
