% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{visualize.power.grid}
\alias{visualize.power.grid}
\title{Plotting function to visualize calculated power as heatmap, splitted into the
separate components of expression probability, power and detection power}
\usage{
visualize.power.grid(
  powerDf,
  var.axis1 = "sampleSize",
  var.axis2 = "totalCells"
)
}
\arguments{
\item{powerDf}{Data frame with power calculation for different parameter combinations}

\item{var.axis1}{Variable which should be plotted at the x axis
(possible values: sampleSize, totalCells and readDepth)}

\item{var.axis2}{Variable which should be plotted at the y axis
(possible values: sampleSize, totalCells and readDepth)}
}
\value{
ggplot object with a heatmap of the two power components plus overall power
}
\description{
Plotting function to visualize calculated power as heatmap, splitted into the
separate components of expression probability, power and detection power
}
