% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{calculate.probabilities}
\alias{calculate.probabilities}
\title{Help function to calculate expression probability, power and detection power
for a given gamma distribution plus additional parameters}
\usage{
calculate.probabilities(
  nSamples,
  ctCells,
  type,
  ref.study,
  ref.study.name,
  gamma.parameters,
  disp.fun,
  min.UMI.counts,
  perc.indiv.expr,
  nGenes,
  samplingMethod,
  sign.threshold,
  MTmethod,
  returnResultsDetailed
)
}
\arguments{
\item{nSamples}{Sample size}

\item{ctCells}{Number of cells of the target cell type}

\item{type}{(eqtl/de) study}

\item{ref.study}{Data frame with reference studies to be used for expression ranks and effect sizes
(required columns: name (study name), rank (expression rank), FoldChange (DE study) /Rsq (eQTL study))}

\item{ref.study.name}{Name of the reference study. Will be checked in the ref.study data frame for it (as column name).}

\item{gamma.parameters}{Data frame with gamma parameters, filtered for the correct cell type
(required columns: ct (cell type), s1, r1, s2, r2, p1, p2/p3 (gamma parameters for both components))}

\item{disp.fun}{Function to fit the dispersion parameter dependent on the mean,
filtered for the correct cell type (required columns: ct (cell type),
asymptDisp, extraPois (both from taken from DEseq))}

\item{min.UMI.counts}{Expression defition parameter: more than is number of UMI counts for each
gene per individual and cell type is required to defined it as expressed in one individual}

\item{perc.indiv.expr}{Expression defition parameter: percentage of individuals that need to have this gene expressed
to define it as globally expressed}

\item{nGenes}{Number of genes to simulate (should match the number of genes used for the fitting)}

\item{samplingMethod}{Approach to sample the gene mean values (either taking quantiles or random sampling)}

\item{sign.threshold}{Significance threshold}

\item{MTmethod}{Multiple testing correction method
(possible options: "Bonferroni","FDR","none")}

\item{returnResultsDetailed}{If true, return not only summary data frame, but additional list with exact probability vectors}
}
\value{
Power to detect the DE/eQTL genes from the reference study in a single cell experiment with these parameters
}
\description{
Help function to calculate expression probability, power and detection power
for a given gamma distribution plus additional parameters
}
