% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_fit.R
\name{estimate.exp.prob.values}
\alias{estimate.exp.prob.values}
\title{Estimate gene expression probability based on mean and dispersion values}
\usage{
estimate.exp.prob.values(
  mu,
  size,
  nCellsCt,
  nSamples,
  min.counts = 10,
  perc.indiv = 0.5
)
}
\arguments{
\item{mu}{Estimated mean value in sc data per gene (vector)}

\item{size}{Estimated size parameter in sc data from the negative binomial fit (1/dispersion parameter), also per gene (vector)}

\item{nCellsCt}{Mean number of cells per person and cell type}

\item{nSamples}{Total sample size}

\item{min.counts}{More than is number of UMI counts for each gene per person and cell type is required to defined it as expressed in one individual}

\item{perc.indiv}{Percentage of individuals that need to have this gene expressed to define it as globally expressed}
}
\value{
Vector with expression probabilities for each gene
}
\description{
This function estimates the expression probability of each gene in pseudobulk
with a certain cutoff of more than min.counts UMI counts, based on the mean and
the disperion value of each gene
}
