% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_fit.R
\name{mixed.gamma.estimation}
\alias{mixed.gamma.estimation}
\title{Estimate the mixed gamma distribution of the mean values}
\usage{
mixed.gamma.estimation(
  mean.vals,
  censoredPoint = NULL,
  num.genes.kept = 21000,
  return.df = TRUE
)
}
\arguments{
\item{mean.vals}{Vector with all normalized mean values (estimated using nbinom.estimation)}

\item{censoredPoint}{Censoring point for left censored gamma distributions (if not set,
the minimal positive value will be chosen)}

\item{num.genes.kept}{Total number of genes used for the fit (remove additional genes with a mean of 0)}

\item{return.df}{If true return data frame with fitted parameters, otherwise an object of class EMResult
(containing also the loglikelihoods)}
}
\value{
Data frame with the six parameters describing the gamma mixed distributions
}
\description{
The estimated distribution will be a mixture of a zero component and two left zensored gamma distributions,
using an em fitting procedure implemented in the package. It is recommended to reduce the total number of
fitted genes (often in the count matrix a huge fraction of zero genes) using the parameter num.genes.kept.
}
