% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_fit.R
\name{nbinom.estimation}
\alias{nbinom.estimation}
\title{Estimate the mean and dispersion parameter for each gene}
\usage{
nbinom.estimation(counts.ct, sizeFactors = "standard")
}
\arguments{
\item{counts.ct}{Count matrix as in import for DEseq}

\item{sizeFactors}{Use to different size factor normalization methods, either the "standard"
method from DEseq or "poscounts" of DESeq2}
}
\value{
list with three data frame containing the normalized mean values,
the dispersion parameters and the parameters for the mean-dispersion function
estimated from DESeq
}
\description{
Estimate the mean and dispersion parameter for each gene
}
