% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{optimize.constant.budget.restrictedDoublets}
\alias{optimize.constant.budget.restrictedDoublets}
\title{Optimizing cost parameters to maximize detection power for a given budget and 10X design}
\usage{
optimize.constant.budget.restrictedDoublets(
  totalBudget,
  readDepthRange,
  cellPersRange,
  costKit,
  costFlowCell,
  readsPerFlowcell,
  ct.freq,
  type,
  ref.study,
  ref.study.name,
  cellsPerLane,
  read.umi.fit,
  gamma.mixed.fits,
  gamma.probs,
  ct,
  disp.fun.param,
  mappingEfficiency = 0.8,
  multipletRate = 7.67e-06,
  multipletFactor = 1.82,
  min.UMI.counts = 10,
  perc.indiv.expr = 0.5,
  nGenes = 21000,
  samplingMethod = "quantiles"
)
}
\arguments{
\item{totalBudget}{Overall experimental budget}

\item{readDepthRange}{Range of read depth values that should be tested (vector)}

\item{cellPersRange}{Range of cells per person that should be tested (vector)}

\item{costKit}{Cost for one 10X kit}

\item{costFlowCell}{Cost of one flow cells for sequencing}

\item{readsPerFlowcell}{Number reads that can be sequenced with one flow cell}

\item{ct.freq}{Frequency of the cell type of interest}

\item{type}{(eqtl/de) study}

\item{ref.study}{Data frame with reference studies to be used for effect sizes and ranks (required columns:)}

\item{ref.study.name}{Name of the reference study. Will be checked in the ref.study data frame for it (as column name).}

\item{cellsPerLane}{Maximal number of cells per 10X lane}

\item{read.umi.fit}{Data frame for fitting the mean UMI counts per cell depending on the mean readds per cell (required columns: )}

\item{gamma.mixed.fits}{Data frame with gamma mixed fit parameters for each cell type (required columns: )}

\item{gamma.probs}{Data frame with probability parameter of the gamma distributions
(required columns:ct (cell type), pi.c1 (probability of component 1))}

\item{ct}{Cell type of interest (name from the gamma mixed models)}

\item{disp.fun.param}{Function to fit the dispersion parameter dependent on the mean (required columns:)}

\item{mappingEfficiency}{Fraction of reads successfully mapped to the transcriptome in the end (need to be between 1-0)}

\item{multipletRate}{Expected increase in multiplets for additional cell in the lane}

\item{multipletFactor}{Expected read proportion of multiplet cells vs singlet cells}

\item{min.UMI.counts}{Expression defition parameter: more than is number of UMI counts for each
gene per person and cell type is required to defined it as expressed in one individual}

\item{perc.indiv.expr}{Expression defition parameter: percentage of individuals that need to have this gene expressed
to define it as globally expressed}

\item{nGenes}{Number of genes to simulate (should match the number of genes used for the fitting)}

\item{samplingMethod}{Approach to sample the gene mean values (either taking quantiles or random sampling)}

\item{totalCost}{Experimental budget}
}
\description{
This function determines the optimal parameter combination for a given budget.
The optimal combination is thereby the one with the highest detection power.
The parameters are checked for a range of read depths and cells per person values,
while the sample size is defined uniquely given the other two parameters and
the overall budget.
}
\examples{
optimize.constant.budget(10000,seq(1000,10000,by=1000),seq(1000,10000,by=1000),
5600,14032,4100*10^6,0.2,"de",de.ref.study,"Blueprint (CLL) iCLL-mCLL",8,
read.umi.fit,gamma.mixed.fits,"CD4 T cells",disp.fun.param)

}
