% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power.eqtl}
\alias{power.eqtl}
\title{Power calculation for an eQTL gene using the F-test}
\usage{
power.eqtl(heritability, sig.level, nSamples)
}
\arguments{
\item{heritability}{Heritability of the trait}

\item{sig.level}{Significane threshold}

\item{nSamples}{Sample size}
}
\value{
Power to detect the eQTL gene
}
\description{
This function calculates the power to detect an eQTL gene.
}
