% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power.general.withDoublets}
\alias{power.general.withDoublets}
\title{Power calculation for a DE/eQTL study  with 10X design (with a restricted number of individuals per lane)}
\usage{
power.general.withDoublets(
  nSamples,
  nCells,
  readDepth,
  ct.freq,
  type,
  ref.study,
  ref.study.name,
  samplesPerLane,
  read.umi.fit,
  gamma.mixed.fits,
  ct,
  disp.fun.param,
  mappingEfficiency = 0.8,
  multipletRate = 7.67e-06,
  multipletFactor = 1.82,
  min.UMI.counts = 10,
  perc.indiv.expr = 0.5,
  nGenes = 21000,
  samplingMethod = "quantiles",
  multipletRateGrowth = "linear"
)
}
\arguments{
\item{nSamples}{Sample size}

\item{nCells}{Number of cells per individual}

\item{readDepth}{Target read depth per cell}

\item{ct.freq}{Frequency of the cell type of interest}

\item{type}{(eqtl/de) study}

\item{ref.study}{Data frame with reference studies to be used for expression ranks and effect sizes
(required columns: name (study name), rank (expression rank), FoldChange (DE study) /Rsq (eQTL study))}

\item{ref.study.name}{Name of the reference study. Will be checked in the ref.study data frame for it (as column name).}

\item{samplesPerLane}{Maximal number of individuals per 10X lane}

\item{read.umi.fit}{Data frame for fitting the mean UMI counts per cell depending on the mean readds per cell
(required columns: intercept, reads (slope))}

\item{gamma.mixed.fits}{Data frame with gamma mixed fit parameters for each cell type
(required columns: parameter, ct (cell type), intercept, meanUMI (slope))}

\item{ct}{Cell type of interest (name from the gamma mixed models)}

\item{disp.fun.param}{Function to fit the dispersion parameter dependent on the mean
(required columns: ct (cell type), asymptDisp, extraPois (both from taken from DEseq))}

\item{mappingEfficiency}{Fraction of reads successfully mapped to the transcriptome in the end (need to be between 1-0)}

\item{multipletRate}{Expected increase in multiplets for additional cell in the lane}

\item{multipletFactor}{Expected read proportion of multiplet cells vs singlet cells}

\item{min.UMI.counts}{Expression defition parameter: more than is number of UMI counts for each
gene per individual and cell type is required to defined it as expressed in one individual}

\item{perc.indiv.expr}{Expression defition parameter: percentage of individuals that need to have this gene expressed
to define it as globally expressed}

\item{nGenes}{Number of genes to simulate (should match the number of genes used for the fitting)}

\item{samplingMethod}{Approach to sample the gene mean values (either taking quantiles or random sampling)}

\item{multipletRateGrowth}{Development of multiplet rate with increasing number of cells per lane, "linear" if overloading should be
modeled explicitly, otherwise "constant". The default value for the parameter multipletRate is matching the option "linear".}
}
\value{
Power to detect the DE/eQTL genes from the reference study in a single cell experiment with these parameters
}
\description{
This function to calculate the detection power for a DE or eQTL study, given DE/eQTL genes from a reference study,
in a single cell 10X RNAseq study. The power depends on the cost determining parameter of sample size, number of cells
per individual and read depth.
}
\examples{
power.general.withDoublets(83,1000,1000,0.2,"de",de.ref.study, "Blueprint (CLL) iCLL-mCLL",
8,read.umi.fit,gamma.mixed.fits,"CD4 T cells",disp.fun.param)

}
