% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power.smartseq}
\alias{power.smartseq}
\title{Power calculation for a DE/eQTL study  with Smart-seq design}
\usage{
power.smartseq(
  nSamples,
  nCells,
  readDepth,
  ct.freq,
  type,
  ref.study,
  ref.study.name,
  gamma.mixed.fits,
  ct,
  disp.fun.param,
  mappingEfficiency = 0.8,
  multipletFraction = 0,
  multipletFactor = 1.82,
  min.norm.count = 10,
  perc.indiv.expr = 0.5,
  nGenes = 21000,
  samplingMethod = "quantiles"
)
}
\arguments{
\item{nSamples}{Sample size}

\item{nCells}{Number of cells per person}

\item{readDepth}{Target read depth per cell}

\item{ct.freq}{Frequency of the cell type of interest}

\item{type}{(eqtl/de) study}

\item{ref.study}{Data frame with reference studies to be used for expression ranks and effect sizes
(required columns: rank, FoldChange (DE study) /Rsq (eQTL study))}

\item{ref.study.name}{Name of the reference study. Will be checked in the ref.study data frame for it (as column name).}

\item{gamma.mixed.fits}{Data frame with gamma mixed fit parameters for each cell type (required columns: )}

\item{ct}{Cell type of interest (name from the gamma mixed models)}

\item{disp.fun.param}{Function to fit the dispersion parameter dependent on the mean (required columns:)}

\item{mappingEfficiency}{Fraction of reads successfully mapped to the transcriptome in the end (need to be between 1-0)}

\item{multipletFraction}{Multiplet fraction in the experiment as a constant factor}

\item{multipletFactor}{Expected read proportion of multiplet cells vs singlet cells}

\item{min.norm.count}{Expression defition parameter: more than is number of counts per kilobase transcript for each
gene per person and cell type is required to defined it as expressed in one individual}

\item{perc.indiv.expr}{Expression defition parameter: percentage of individuals that need to have this gene expressed
to define it as globally expressed}

\item{nGenes}{Number of genes to simulate (should match the number of genes used for the fitting)}

\item{samplingMethod}{Approach to sample the gene mean values (either taking quantiles or random sampling)}
}
\value{
Power to detect the DE/eQTL genes from the reference study in a single cell experiment with these parameters
}
\description{
This function to calculate the detection power for a DE or eQTL study, given DE/eQTL genes from a reference study,
in a single cell Smart-seq RNAseq study. The power depends on the cost determining parameter of sample size, number of cells
per person and read depth.
}
