% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{sampleSizeBudgetCalculation}
\alias{sampleSizeBudgetCalculation}
\title{Estimate possible sample size depending on the total cost and the other parameters for 10X design}
\usage{
sampleSizeBudgetCalculation(
  nCells,
  readDepth,
  totalCost,
  costKit,
  samplesPerLane,
  costFlowCell,
  readsPerFlowcell,
  reactionsPerKit = 6
)
}
\arguments{
\item{nCells}{Cells per individual}

\item{readDepth}{Read depth per cell}

\item{totalCost}{Experimental budget}

\item{costKit}{Cost for one 10X kit}

\item{samplesPerLane}{Number of individuals sequenced per lane}

\item{costFlowCell}{Cost of one flow cells for sequencing}

\item{readsPerFlowcell}{Number reads that can be sequenced with one flow cell}

\item{reactionsPerKit}{Reactions (=lanes) per kit, defines the total
number of tested individuals per kit}
}
\value{
Number of samples that can be measured with this budget and other parameters
}
\description{
A balanced design with two classes is assumed for the sample size calculation.
}
