% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{sampleSizeBudgetCalculation.restrictedDoublets}
\alias{sampleSizeBudgetCalculation.restrictedDoublets}
\title{Estimate possible sample size depending on the total cost and the other parameters
(with a restricted number of cells per lane)}
\usage{
sampleSizeBudgetCalculation.restrictedDoublets(
  nCells,
  readDepth,
  totalCost,
  costKit,
  cellsPerLane,
  costFlowCell,
  readsPerFlowcell
)
}
\arguments{
\item{nCells}{Cells per person}

\item{readDepth}{Read depth per cell}

\item{totalCost}{Experimental budget}

\item{costKit}{Cost for one 10X kit}

\item{cellsPerLane}{Number of cells sequenced per lane}

\item{costFlowCell}{Cost of one flow cells for sequencing}

\item{readsPerFlowcell}{Number reads that can be sequenced with one flow cell}
}
\value{
Number of samples that can be sampled with this budget and other parameters
}
\description{
Variant of sampleSizeBudgetCalculation, which restricts the cells per lane instead the persons
per lane.
}
