% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{size.estimates}
\alias{size.estimates}
\title{Precalcuated size estimates for eQTL power simulation}
\format{Data frame with precalculated size estimates (depending on Rsq, af and mean)}
\usage{
data(sizeEstimates)

data(sizeEstimates)
}
\description{
Precalculated size estimates for eQTL power simulation
(using the function estimate.size.simulation) to speed simulation

Precalculated p-values from eQTL simulation to speed calculation time
}
\examples{
data(sizeEstimates)
head(size.estimates)
data(simEqtlPvals)
head(simEqtlPvals)
}
\keyword{datasets}
