% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_fit.R
\name{umi.read.relation}
\alias{umi.read.relation}
\title{Estimate logarithmic relationship between mean UMI counts and mean mapped read counts per cell}
\usage{
umi.read.relation(read.umis)
}
\arguments{
\item{read.umis}{Data.frame with column for mean UMI values (mean.umi) and
column for mean transcriptome mapped reads (transcriptome.mapped.reads)}
}
\value{
Parameters of logarithmic fit between reads and UMIs
}
\description{
Estimate logarithmic relationship between mean UMI counts and mean mapped read counts per cell
}
